Turn off extension by default
Summary
-----

This commit turns off the extension unless a metadata tag is explicitly
added in the notebook itself. This prevents untested notebooks from
increasing in size, especically in the Data 8 setting where we only want
to get the history of predetermined notebooks. Turning the extension on
for all notebooks can result in students having big notebooks for other
classes too which is bad.

How to test
-----

1. Create a new notebook.
2. Run a cell a couple times and verify that the history is not saved.
3. Add `timetravel: { enabled: true }` to the notebook metadata.
4. Verify that the notebook does record history.